Correct the WiFi Beacon OUI/CID identifier
The early drafts of the ASD-STAN Direct Remote ID
specification were using a OUI for WiFi Beacon that
belongs to the drone manufacturer Parrot. The final
version of the specification is using a CID instead,
belonging to ASD-STAN.

Add also verification of the Vendor Type, before
accepting WiFi Beacon data.

Signed-off-by: Soren Friis <soren.friis@intel.com>