Move authentication data to the end of the log file
It is often not present but adds a lot of columns in the log file,
which can make it hard to find the self ID, System and Operator ID data.

Also skip any possible (unsupported) multi-messages of the same type,
to ensure the log layout is consistent and logs all other messages.

Signed-off-by: Soren Friis <soren.friis@intel.com>