Support multiple Basic ID messages
New addition in the ASTM specification v1.1

Some regions require the ability to send more than one
Basic ID message to relay e.g. both the serial number
and a CAA registration ID.

Currently only up-to two types of Basic ID messages can
be received (and logged to the csv file). If a third type
is received, it will be silently ignored.

Signed-off-by: Soren Friis <soren.friis@intel.com>